[temp-workmem P4-followup] segment-list scan (class-A) + qmgr_segment_pos_read + interim materialize (guard inert; parity-preserving)
Wires the class-A/A-prime consumer mechanisms over coordinator-owned raw-fd segments;
INERT while the raw-fd write guard is FALSE (no raw-fd segments produced -> develop path
byte-identical), preparing for the P5 raw-fd activation:
- QMGR_SEGMENT / QMGR_SEGMENT_LIST / QMGR_SEGMENT_LIST_SCAN + qmgr_segment_list_* APIs:
  coordinator-owned ordered set of per-worker raw-fd segments with a class-A SEQUENTIAL scan.
  merge_list_ids (px_scan_result_handler) builds a segment list + uses the segment scan ONLY when
  every nonempty worker result is a raw-fd segment; otherwise falls back to the develop
  qfile_connect_list/copy path. Class-A segment path uses NO materialize.
- qmgr_segment_pos_read: QFILE positioned accessor that reads a raw-fd tuple coordinate via
  temp_page_store::rawfd_pos_read (read-cache backed). qfile_jump_scan_tuple_position now routes
  raw-fd coordinates here (replacing the P4-foundational inert stub). Unreachable in product
  execution while guard FALSE (raw-fd coords are only produced when raw-fd writes are enabled);
  exercised by the debug SEGPOS self-test against a directly-created raw-fd segment.
- qmgr_materialize_to_pgbuf + A-prime INTERIM-materialize hook: qfile_open_list_scan materializes
  raw-fd-backed list chains to a real-VPID pgbuf list for A-prime candidates (so P4 is
  parity-correct without switching the six A-prime operators to segment-position yet -- that is P5).
- Advisory #1 fixed: raw_fd_segment_id widened to UINT64 in QFILE_TUPLE_POSITION / _DB / SIMPLE_POS
  (matches temp_page_store::raw_fd_page_coordinate), store/restore/size constants updated.

Verified (debug, tde_selftest): build clean; SEGPOS/RAWFD/TEMPMOVE self-tests result=0 (SEGPOS now
also round-trips qmgr_segment_pos_read against a real raw-fd segment); CONNECT BY 5-level=1..5 +
200-level=200 (no hang); ORDER BY correct; guard FALSE (LEADER_VERIFIED_ENABLE_RAW_FD_WRITES=false).
P5 switches the six A-prime operators to segment-position + installs class-B materialize-at-sink +
flips the guard (gated on the full P1b safety verification).

+++ src/query/query_hash_scan.c
@@ -491,8 +491,9 @@ qdata_print_hash_scan_entry (THREAD_ENTRY * thread_p, FILE * fp, const void *dat
     {
       if (qfile_tuple_simple_pos_is_raw_fd (data_p->pos))
 	{
-	  fprintf (fp, "raw_fd_segment_id = [%u]  page_index = [%d]  tuple_offset = [%d]",
-		   data_p->pos->raw_fd_segment_id, data_p->pos->page_index, data_p->pos->tuple_offset);
+	  fprintf (fp, "raw_fd_segment_id = [%llu]  page_index = [%d]  tuple_offset = [%d]",
+		   (unsigned long long) data_p->pos->raw_fd_segment_id, data_p->pos->page_index,
+		   data_p->pos->tuple_offset);
 	}
       else
 	{
